create activedoptree and workerattr modules
