+++ settings.cfg
@@ -8,5 +8,34 @@ ACCOUNTS = {
 	'JoeAnnotator': 'example',
 	'SecondAnnotator': 'example',
 	}
+PUNCT_ESCAPING = [{"istring" : "(", "ptree_label": "LRB-p", "ptree_token": "-LRB-", "ctree_punct": "("},
+				  {"istring" : ")", "ptree_label": "RRB-p", "ptree_token": "-RRB-", "ctree_punct": ")"},
+				  {"istring" : "-", "ptree_label": "HYPH-p", "ptree_token": "-", "ctree_punct": "-"}]
+# punctuation labels based on the Penn Treebank tagset [from nltk.help.upenn_tagset()]
+PUNCT_TAGS = { 
+	# opening parenthesis
+	"-LRB-": "LRB", "[" : "LRB", "{": "LRB",
+	# closing parenthesis
+	"-RRB-": "RRB", "]": "RRB", "}": "RRB",
+	# sentence terminator
+	".": ".", "!": ".", "?": ".",
+	# colon or ellipsis
+	":": ":", ";": ":", "...": ":",
+	# comma
+	",": ",",
+	# dash
+	"-": "HYPH", "--": "HYPH",
+	# opening quotation mark
+	"``": "``", "`": "``",
+	# closing quotation mark
+	"''": "''", "'": "''"
+}
+SYMBOL_TAG = "*"
+AMBIG_SYM = {
+	"$", "#", "%",	# typically N-Head
+	"@",	# typically P-Head
+	"&", "-", "/"	# typically Coordinator-Marker
+}
+INITIAL_PUNCT_LABELS = {'LRB-p', '[-p', '{-p'}
 FUNCTIONTAGWHITELIST = ['Subj+PCComp', 'Obj+PCComp', 'Head+PCComp', 'p']
 CGELVALIDATE = True
